♻️ [Refactoring]: #20 ObjectStreamExtractor モジュールを責務ごとに分割
469行の単一ファイルを5つのサブモジュールに分割し、命名・公開API境界を改善:
- types.ts: インタフェース定義
- dict/: ObjectStreamDict (validate + ObjectStreamDictInfo、内部モジュール)
- header/: ObjectStreamHeader (parse + ObjectStreamHeaderEntry、公開型)
- flate-decompressor/: createFlateDecompressor + DEFAULT定数
- extractor/: ObjectStreamExtractor クラス本体

公開APIから ValidatedStreamDict/validateStreamDict/parseHeader を除去し、
ObjectStreamHeader.parse + ObjectStreamHeaderEntry に置き換え。
テストヘルパーを共通化し、テストファイルを対象モジュールと同フォルダに再配置。
